FIREFLY-1463:  Add CorsFilter - apply CorsFilter to font files - moved cors logic to this class
